requirements.txt
fastapi
uvicorn
python-multipart
requests
beautifulsoup4
plotly
streamlit
pandas
numpy

# 🔐 AUTH (PINNED)
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
python-dotenv

# 🗃️ DATABASE
psycopg2-binary
